% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorize_data.R
\name{factorize_character_vector}
\alias{factorize_character_vector}
\title{Turn character vector into factor}
\usage{
factorize_character_vector(data_vector, custom_labels = NULL)
}
\arguments{
\item{data_vector}{Vector with character data.}

\item{custom_labels}{Custom labels named vector, names=levels, value=labels}
}
\value{
Factorized vector
}
\description{
Turn character vector into factor
}
\examples{
factorize_character_vector(c("First", "Second", "Third"))
}
