% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorize_data.R
\name{factorize_nicely_dataframe}
\alias{factorize_nicely_dataframe}
\title{Heuristic factorization for all columns of a data frame}
\usage{
factorize_nicely_dataframe(data_frame)
}
\arguments{
\item{data_frame}{A data frame.}
}
\value{
A data frame with all columns converted to factors.
}
\description{
Heuristic factorization for all columns of a data frame
}
\examples{
df <- data.frame(x = runif(20), y = rep(c(TRUE, FALSE, TRUE, TRUE), 5))
factorize_nicely_dataframe(df)
}
