% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorize_data.R
\name{factorize_nicely_vector}
\alias{factorize_nicely_vector}
\title{Heuristic factorization for a single vector}
\usage{
factorize_nicely_vector(data_vector)
}
\arguments{
\item{data_vector}{A vector (numeric, logical, or character).}
}
\value{
A factor (ordered for numeric inputs with many distinct values).
}
\description{
Heuristic factorization for a single vector
}
\examples{
factorize_nicely_vector(c("a", "b", "a"))
}
