% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_data}
\alias{load_data}
\title{Loads data for the use of MLmorph}
\usage{
load_data(data_path)
}
\arguments{
\item{data_path}{File path of the data}
}
\value{
A data.frame with imported data
}
\description{
Loads data for the use of MLmorph
}
\examples{
xlsxFile <- system.file("extdata", "load_data_test.xlsx", package = "openxlsx")
df1 <- load_data(xlsxFile)

csvFile <- system.file("extdata", "load_data_test.csv", package = "openxlsx")
df2 <- load_data(csvFile)

jsonFile <- system.file("extdata", "load_data_test.json", package = "openxlsx")
df3 <- load_data(jsonFile)
}
