% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLmorph.R
\name{MLmorph_live}
\alias{MLmorph_live}
\title{Launch MLmorph from the source tree (development helper)}
\usage{
MLmorph_live()
}
\description{
Launch MLmorph from the source tree (development helper)
}
\examples{
\dontrun{
# From package root during development
MLmorph_live()
}
}
\keyword{internal}
