% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_model.R
\name{create_rf_model}
\alias{create_rf_model}
\title{Create random forests model}
\usage{
create_rf_model(
  data,
  dependent = colnames(data)[ncol(data)],
  independent = setdiff(colnames(data), dependent),
  train_validate_split = 0.8,
  ...
)
}
\arguments{
\item{data}{The data.frame with data used for the model estimation}

\item{dependent}{Predictor variables names}

\item{independent}{The predicted variable name}

\item{train_validate_split}{The proportion of data used for training, vs validation}

\item{...}{Other arguments passed to the randomForest::randomForest() function}
}
\value{
Named list, with: model, variables_importance, model_performance_on_test
}
\description{
Create random forests model
}
\examples{
create_rf_model(data.frame(a = factorize_numeric_vector(runif(100, 10, 20), breaks_no = 3),
b = factorize_numeric_vector(runif(100, 1, 2), breaks_no = 5),
c = factorize_numeric_vector(runif(100, 4, 5), breaks_no = 2)))
}
