% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorize_data.R
\name{factorize_binary_vector}
\alias{factorize_binary_vector}
\title{Turn binary vector into factor}
\usage{
factorize_binary_vector(data_vector, custom_labels = NULL)
}
\arguments{
\item{data_vector}{Vector with binary data.}

\item{custom_labels}{Custom labels vector: first for TRUE, second for FALSE.}
}
\value{
Factorized vector
}
\description{
Turn binary vector into factor
}
\examples{
factorize_binary_vector(c(TRUE, FALSE, TRUE))
}
