% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorize_data.R
\name{factorize_character_vector}
\alias{factorize_character_vector}
\title{Turn character vector into a factor}
\usage{
factorize_character_vector(data_vector, custom_labels = NULL)
}
\arguments{
\item{data_vector}{Character vector.}

\item{custom_labels}{Optional named character vector where names are original values and values are labels.}
}
\value{
A factor with labeled levels.
}
\description{
Turn character vector into a factor
}
\examples{
factorize_character_vector(c("A First", "B Second", "C Third"))
}
