% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorize_data.R
\name{factorize_numeric_vector}
\alias{factorize_numeric_vector}
\title{Turn numeric vector into ordered factor}
\usage{
factorize_numeric_vector(
  data_vector,
  method = c("equal_bins", "equal_distance", "custom_breaks"),
  breaks_no = 5,
  custom_breaks = NULL,
  custom_labels = NULL
)
}
\arguments{
\item{data_vector}{Vector with numeric data.}

\item{method}{Method to create breaks, "equal_bins" by default.}

\item{breaks_no}{Number of breaks to create, 5L by default.}

\item{custom_breaks}{Custom breaks to supply, as a named numeric vector.}

\item{custom_labels}{Custom labels to supply, as a named numeric vector.}
}
\value{
Factorized vector
}
\description{
Turn numeric vector into ordered factor
}
\examples{
factorize_numeric_vector(runif(50))
}
