% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLmorph-package.R
\docType{package}
\name{MLmorph-package}
\alias{MLmorph}
\alias{MLmorph-package}
\title{MLmorph: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
