% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLmorph.R
\name{MLmorph_dependencies}
\alias{MLmorph_dependencies}
\title{Iterate through app libraries and functions}
\usage{
MLmorph_dependencies()
}
\description{
This function references packages and functions used in the MLmorph Shiny app,
located in \verb{inst/}.
}
\keyword{internal}
